+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# Next Up
+
+- Ensnaring Strike fixes from the wonderful @elwin1410
+
 # 5.1.6
 
 - Some effects, noteably Magic Weapon, would not remove the magic property from the weapon when the effect ended.